requirements.txt
numpy
opencv-python
tqdm
openpyxl
Pillow

# dev
matplotlib
jupyter
jupyterlab
pyinstaller
pylint

# python >= 3.7
